+++ changelog.md
@@ -1,6 +1,17 @@
-### GEOGLAM Map Change Log
+### GEOGLAM RAPP Change Log
+
+#### 2016-03-18
+
+* Layers renamed and rearranged
+* Data61 logo replaces NICTA logo on about page
+* Start map with 2D view of whole world
+* Name of map changed from GEOGLAM Map to GEOGLAM RAPP
+* Logo rearranged: GEOGLAM RAPP bigger and to the left, with Data61 on the right
+* Added site disclaimer to declare GEOGLAM RAPP as under development
 
 #### 2016-03-01
 
 * First release
+
+